+++ src/main/java/it/unibo/superpeach/blocks/MapFixedBlock.java
@@ -4,7 +4,7 @@ import java.awt.Graphics;
 
 public class MapFixedBlock extends Block{
 
-    public MapFixedBlock(double x, double y, double width, double height, int scale, BlockType type) {
+    public MapFixedBlock(int x, int y, int width, int height, int scale, BlockType type) {
         super(x, y, width, height, scale);
         setType(type);
     }
@@ -18,17 +18,17 @@ public class MapFixedBlock extends Block{
         switch (getType()) {
             case TERRAIN:
                 setSprites(getTexturer().getBricksCastle());
-                g.drawImage(getSprites()[0], (int)getX(), (int)getY(), (int)getWidth(), (int)getHeight(), null);
+                g.drawImage(getSprites()[0], getX(), getY(), getWidth(), getHeight(), null);
                 break;
 
             case LUCKY:
                 setSprites(getTexturer().getLucky());
-                g.drawImage(getSprites()[0], (int)getX(), (int)getY(), (int)getWidth(), (int)getHeight(), null);
+                g.drawImage(getSprites()[0], getX(), getY(), getWidth(), getHeight(), null);
                 break;
 
             case BRICK:
                 setSprites(getTexturer().getBricksCastle());
-                g.drawImage(getSprites()[2], (int)getX(), (int)getY(), (int)getWidth(), (int)getHeight(), null);
+                g.drawImage(getSprites()[2], getX(), getY(), getWidth(), getHeight(), null);
                 break;
         }
     }
